feat: 결제 개발 및 주문 관련 api 개발

+++ src/main/java/com/example/farmfarm_refact/entity/OrderDetailEntity.java
@@ -22,7 +22,7 @@ public class OrderDetailEntity {
     @Id
     @GeneratedValue
     @Column(name="od_id")
-    private Integer odId;
+    private Long odId;
 
     private int quantity;
